requirements-venv_auto_ml_pycaret.txt
annotated-types==0.7.0
ansi2html==1.9.2
aplr==10.20.1
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
category_encoders==2.7.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
choreographer==1.2.1
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
colour==0.1.5
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.3
dacite==1.9.2
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_auth==2.3.0
dash_cytoscape==1.0.2
debugpy==1.8.19
decorator==5.2.1
deprecation==2.1.0
dill==0.4.0
dtreeviz==2.2.2
executing==2.2.1
explainerdashboard==0.5.1
fairlearn==0.7.0
fastjsonschema==2.21.2
filetype==1.2.0
Flask==3.0.3
Flask-WTF==1.2.2
flask_simplelogin==0.2.0
fonttools==4.61.1
gevent==25.9.1
graphviz==0.21
greenlet==3.3.0
idna==3.11
ImageHash==4.3.2
imbalanced-learn==0.14.1
importlib_metadata==8.7.1
iniconfig==2.3.0
interpret==0.7.3
interpret-core==0.7.3
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.3.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-dash==0.4.2
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kaleido==1.2.0
kiwisolver==1.4.9
lightgbm==4.6.0
llvmlite==0.45.1
logistro==2.0.1
MarkupSafe==3.0.3
matplotlib==3.7.5
matplotlib-inline==0.2.1
minify_html==0.18.1
multimethod==1.12
multiprocess==0.70.18
narwhals==2.14.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
numba==0.62.1
numpy==1.26.4
orjson==3.11.5
oyaml==1.0
packaging==25.0
pandas==2.1.4
parso==0.8.5
patsy==1.0.2
phik==0.12.5
pillow==12.0.0
platformdirs==4.5.1
plotly==5.24.1
plotly-resampler==0.11.0
pluggy==1.6.0
pmdarima==2.0.4
prompt_toolkit==3.0.52
psutil==7.2.0
pure_eval==0.2.3
puremagic==1.30
pycaret==3.3.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pynndescent==0.5.13
pyod==2.0.6
pyparsing==3.3.1
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.9.0
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
retrying==1.4.2
rpds-py==0.30.0
SALib==1.5.1
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
shap==0.44.1
simplejson==3.20.2
six==1.17.0
sklearn-compat==0.1.5
sktime==0.26.0
slicer==0.0.7
stack-data==0.6.3
statsmodels==0.14.6
tbats==1.1.3
tenacity==9.1.2
threadpoolctl==3.6.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
tsdownsample==0.1.4.1
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
umap-learn==0.5.7
urllib3==2.6.2
visions==0.8.1
waitress==3.0.2
wcwidth==0.2.14
Werkzeug==3.0.6
widgetsnbextension==4.0.15
wordcloud==1.9.5
WTForms==3.2.1
xxhash==3.6.0
ydata-profiling==4.18.0
yellowbrick==1.5
zipp==3.23.0
zope.event==6.1
zope.interface==8.1.1
